requirements.txt
streamlit
WeasyPrint
google.generativeai
